\name{HPdgraph-package}
\alias{HPdgraph-package}
\alias{HPdgraph}
\docType{package}
\title{Distributed algorithms for graph analytics}
\description{
\pkg{HPdgraph} provides a few distributed algorithms for graph analytics. It is written based on the infrastructure created in HP-Labs for distributed computing in R.
}

\details{
    \tabular{ll}{
        Package: \tab HPdgraph\cr
        Type: \tab Package\cr
        Version: \tab 0.4.0\cr
        Date: \tab 2014-01-22\cr
    }
 Main Functions:
 \itemize{
  \item {hpdpagerank:} {It is a distributed function for computing pagerank vector of a graph.}
  \item {hpdwhich.max:} {It finds and returns the index of the maximum value stored in a darray.}
 }
}

\author{
Arash Fard <afard@vertica.com>
}
\references{
 \enumerate{
  \item{Using R for Iterative and Incremental Processing. Shivaram Venkataraman, Indrajit Roy, Alvin AuYoung, Rob Schreiber. HotCloud 2012, Boston, USA.}
 }   
}
\keyword{Distributed R}
\keyword{Distributed Graph Analytics}
\keyword{Big Data Analytics}

