\name{HPdgraph-package}
\alias{HPdgraph-package}
\alias{HPdgraph}
\docType{package}
\title{Distributed algorithms for graph analytics}
\description{
\pkg{HPdgraph} provides distributed algorithms for graph analytics. It is written based on the infrastructure created in HP Labs for distributed computing in R.
}

\details{
    \tabular{ll}{
        Package: \tab HPdgraph\cr
        Type: \tab Package\cr
        Version: \tab 1.1.0\cr
        Date: \tab 2015-01-16\cr
    }
 Main Functions:
 \itemize{
  \item {hpdpagerank:} {compute pagerank of a graph in a distributed fashion.}
  \item {hpdwhich.max:} {returns the index of the maximum value stored in a darray.}
 }
}

\author{
HP Vertica Analytics Team <distributedRTeam@external.groups.hp.com>
}
\references{
 \enumerate{
  \item{Using R for Iterative and Incremental Processing. Shivaram Venkataraman, Indrajit Roy, Alvin AuYoung, Rob Schreiber. HotCloud 2012, Boston, USA.}
 }   
}
\keyword{Distributed R}
\keyword{Distributed Graph Analytics}
\keyword{Big Data Analytics}

